requirements.txt
Flask==3.0.3
gunicorn==22.0.0
google-api-python-client==2.126.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
